Move DummyAWG and DummyDAC into main package
